CRUD funcionando normalmente

+++ src/main/java/com/example/locadorabackend/Domain/actor/RequestActor.java
@@ -2,7 +2,7 @@ package com.example.locadorabackend.Domain.actor;
 
 import jakarta.validation.constraints.NotBlank;
 
-public record RequestActor (int id,
+public record RequestActor (Long id,
                             @NotBlank
                             String nome){
